chore: remove feature flag checks (#410)
https://stord.atlassian.net/browse/OMS-2544

We're locked into cloud from now on so there's no point in checking
launch darkly to see if it should be enabled.

The only condition we have left is that any V1 shippers should go to
user management. Once all of those networks are migrated we can delete
all the user management related code too.